fix(render): address review round 1 on the audit batch
The blob write-elision cache was Float32Array against float64 slot
values, so every comparison truncated false and the instance buffer
re-uploaded each frame on exactly the tier the feature targets: the
arrays go Float64 and a behavioral test pins the contract through the
attribute version counter (same slot twice uploads once). The sky
re-prepare branch in prepareZoneAt gates on standardMaterials so the
shadowless tiers, whose HDR stores stay empty by design, stop re-running
prepareZoneSky on every arrival, and its progress callback now completes
after the dome work it awaits. The residency pinned set and the env RT
evict guard also cover envTransition.pending; zoneSkyBiomes reads the
live world's zones like the region list. Review nits: the linear encode
arms refuse an alpha source loudly, the blob painter nulls after
teardown, the prevProgramKey capture comment states its default-key
limitation, loadHdr and releaseHdr share one request normalization, and
the flip-baked regen command is pinned as a greppable doc literal.

+++ src/render/blob_shadows.ts
@@ -65,10 +65,13 @@ export class BlobShadows {
   private readonly texture: THREE.CanvasTexture;
   private readonly matrix = new THREE.Matrix4();
   /** Last uploaded transform per slot, so an unchanged frame uploads nothing. */
-  private readonly lastX = new Float32Array(BLOB_SHADOW_POOL);
-  private readonly lastY = new Float32Array(BLOB_SHADOW_POOL);
-  private readonly lastZ = new Float32Array(BLOB_SHADOW_POOL);
-  private readonly lastScale = new Float32Array(BLOB_SHADOW_POOL);
+  // Float64: the slot values are doubles, and a Float32Array round-trip
+  // truncates them, so every === comparison below would be false for real
+  // world coordinates and the elision would never fire (review round 1).
+  private readonly lastX = new Float64Array(BLOB_SHADOW_POOL);
+  private readonly lastY = new Float64Array(BLOB_SHADOW_POOL);
+  private readonly lastZ = new Float64Array(BLOB_SHADOW_POOL);
+  private readonly lastScale = new Float64Array(BLOB_SHADOW_POOL);
   private count = 0;
   private dirty = false;
